requirements.txt
streamlit>=1.31
PyMuPDF>=1.23.8
python-dotenv>=1.0.1
pydantic>=2.6.1
reportlab>=4.0.7
openai>=1.47.0
